examples: cut -> ksCut, set -> set_key
rename examples, since they result in 'cut' and 'set' binaries
(conflict with unix tool cut and shell builtin set)
