+++ examples/CMakeLists.txt
@@ -22,10 +22,10 @@ foreach (file ${TESTS})
 endforeach (file ${TESTS})
 
 target_link_elektra(cascading elektra-kdb)
-target_link_elektra(cut elektra-kdb)
+target_link_elektra(ksCut elektra-kdb)
 target_link_elektra(kdbget elektra-kdb)
 target_link_elektra(kdbget_error elektra-kdb)
 target_link_elektra(kdbintro elektra-kdb)
 target_link_elektra(kdbopen elektra-kdb)
 target_link_elektra(kdbset elektra-kdb)
-target_link_elektra(set elektra-kdb)
+target_link_elektra(set_key elektra-kdb)
