+++ examples/set_key.c
